+++ src/board/boardbuf.rs
@@ -96,7 +96,7 @@ impl BoardBuf for Board {
     }
 
     fn set(&mut self, bb: Bitboard, pieces: &str) -> Result<&mut Self, String> {
-        if bb.popcount() != pieces.chars().count() {
+        if bb.popcount() != pieces.chars().count() as i32 {
             return Err(format!("Bitboard {} and pieces {} have different counts", bb, pieces));
         }
         for (sq, ch) in bb.iter().zip(pieces.chars()) {
